# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for InnoTaxi user service",
        "title": "InnoTaxi user service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/order": {
            "post": {
                "description": "handler for get order request, allows user to order the taxi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Order Taxi",
                "operationId": "ordertaxi",
                "parameters": [
                    {
                        "description": "order info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OrderData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all is okay",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "description": "handler for get profile request, allows user to check his profile data(name,  phone number, email, rating)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get Profile",
                "operationId": "getprofile",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ProfileData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "handler for delete profile request, allows user to delete his account(user's data will be available for registration for other users)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Delete Profile",
                "operationId": "deleteprofile",
                "parameters": [
                    {
                        "description": "password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeleteData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User is successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "description": "handler for get profile request, allows user to update data(name, phone number, email). You can change any field of this data, but you should input correct data\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update Profile",
                "operationId": "updateprofile",
                "parameters": [
                    {
                        "description": "user's new data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user updated profile",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/rateTrip": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "handler for rateTrip request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating"
                ],
                "summary": "RateTrip",
                "operationId": "rate trip",
                "parameters": [
                    {
                        "description": "Rate the trip",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ratingTrip"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getRate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/rating": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "handler for SignOut request, allows user log out of his account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating"
                ],
                "summary": "GetRating",
                "operationId": "get-rating",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getRate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "description": "handler for SignIn request, allows user to authenticate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignIn",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.InputSignIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/auth/sign-out": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "handler for SignOut request, allows user log out of his account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignOut",
                "responses": {
                    "200": {
                        "description": "user successfully signed-out",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "handler for SignUp request, allows user to register in service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignUp",
                "operationId": "signup",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user successfully created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.DeleteData": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "entity.InputSignIn": {
            "type": "object",
            "required": [
                "password",
                "phone"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entity.ProfileData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
        "entity.UpdateData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                }
            }
        },
        "handler.OrderData": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "taxiType": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "handler.getRate": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
        "handler.ratingTrip": {
            "type": "object",
            "properties": {
                "rating": {
                    "type": "integer"
                }
            }
        }
    }
}